# web/lib/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InjectRequest": {
        "properties": {
          "amount_rupees": {
            "default": 499.0,
            "exclusiveMinimum": 0.0,
            "maximum": 100000.0,
            "title": "Amount Rupees",
            "type": "number"
          },
          "category": {
            "default": "card_expired",
            "title": "Category",
            "type": "string"
          }
        },
        "title": "InjectRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Failed Subscription Recovery Agent \u2014 Razorpay AI Buildathon, Track 03: AI Revenue Recovery",
    "title": "Sahara",
    "version": "1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/control/inject": {
      "post": {
        "description": "Push one failure through the live pipeline and hand back the case it opened.\n\nThe payload is a real Razorpay event body and it enters through `intake()`, so this\nwalks detect \u2192 diagnose \u2192 gate \u2192 decide exactly like a delivery from Razorpay. The\nonly difference is `source`.",
        "operationId": "inject_api_control_inject_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Inject Api Control Inject Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Inject",
        "tags": [
          "control"
        ]
      }
    },
...
    "/api/demo/case": {
      "get": {
        "description": "Poll target for the pipeline page: resolves either a demo_id (a real checkout\norder's subscription_id note) or a case_id (handed back directly by\n/api/control/inject) to whatever case it points at, and returns the same rich\npayload as GET /api/cases/{id} so the page can render the pipeline live.",
        "operationId": "demo_case_api_demo_case_get",
        "parameters": [
          {
            "in": "query",
            "name": "demo_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Demo Id"
            }
          },
          {
            "in": "query",
            "name": "case_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Demo Case Api Demo Case Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Demo Case",
        "tags": [
          "live-demo"
        ]
      }
    },
...
    "/api/demo/force-tick": {
      "post": {
        "description": "Demo-only clock skip: a real `tick()` only executes decisions already past\ntheir real scheduled_for, and RETRY_LATER is genuinely scheduled hours out \u2014 the\nsame wall-clock discipline the live loop runs under (see main.py's tick loop).\n\nWaiting out real hours mid-demo isn't practical, so this does exactly what\nmanually editing scheduled_for and re-running tick would do: it pulls this one\ncase's still-scheduled decision to now, then calls the same executor.tick() the\nbackground loop calls. Nothing about execute_decision's own logic \u2014 invariant\nre-checks, ExecutionRecord writing, attempt reservation \u2014 is bypassed.",
        "operationId": "demo_force_tick_api_demo_force_tick_post",
        "parameters": [
          {
            "in": "query",
            "name": "demo_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Demo Id"
            }
          },
          {
            "in": "query",
            "name": "case_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Demo Force Tick Api Demo Force Tick Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Demo Force Tick",
        "tags": [
          "live-demo"
        ]
      }
    },
...
    "/api/demo/simulate": {
      "post": {
        "description": "Stands in for the next real-world signal a live Sahara would eventually get on\nits own: whether a silent retry actually cleared, or whether the customer paid (or\nignored) an update link. Enters through the identical `webhooks.intake()` every\nreal delivery uses \u2014 `source=\"synthetic\"` is the only difference \u2014 exactly like\n`/api/control/inject`, just addressed at an existing case instead of opening a\nfresh one.",
        "operationId": "demo_simulate_api_demo_simulate_post",
        "parameters": [
          {
            "in": "query",
            "name": "outcome",
            "required": true,
            "schema": {
              "title": "Outcome",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "demo_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Demo Id"
            }
          },
          {
            "in": "query",
            "name": "case_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Demo Simulate Api Demo Simulate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Demo Simulate",
        "tags": [
          "live-demo"
        ]
      }
    },
...
  }
}